More Info Data
Worked on More Info click event by creating variables to hold some of
the information returned by the AJAX request for the lightbox feature
to be constructed later. Also, made some adjustments to the search
input and added a gray font color to differentiate the album title from
the artist.