Revised Sort
Removed sort container and radio buttons, replacing it with links to
sort by newest and oldest on the main page above the search results. In
lightbox mode, increased font size for larger screens and decreased
max-width to 560px. Added more padding to the header at larger screens
sizes. Modified sort function to include the month the album was
released if available (month defaults to 12 if not provided).