User Comment Section
Created the javascript needed to create a simulated user comment
section with user image and username from the Random User Generator API
and comment from the Lorem createText function. Also, moved random user
AJAX request into the success function of the second Spotify AJAX
request to make sure the data was available for adding to Spotify
search results. Finally, the function calls the lead to the creation of
the card and lightbox HTML where moved into the AJAX success function
of the random user request.